ConstantNameGenerator own file
